(bug 46864) Disable 'toc-floated' user option on hevoy
This patch sets the 'toc-floated' user option — used by the
TocTree extension which is installed by all Wikivoyages — to
'false' on the Hebrew project.

Since all the other Wikivoyages seem to be doing just fine
with the option set to 'true' by default, that's what we'll
use. (I'm turning this into an array in case some other
Wikivoyage will want to set this to false; I believe it'll
be easier to do things this way.)

Bug: 46864
Change-Id: I71d31ddc5be4a4b9dd47c0d0bd1404de7c8d7b77