Update patch set 2
Patch Set 2:

This change is intended to replace https://www.wikidata.org/wiki/Special:AbuseFilter/9 for now (there was apparently a discussion on IRC with some devs to disable edits to the Query namespace), but I considered that would be logical to apply it to the talkspace as well. Simply put, and going quite against what you said, the point is that there are currently no queries (the Wikibase extension isn't ready for this as yet), so there is no need to discuss what doesn't exist. Simple put, it's more of a matter of prevention, though this should likely be reverted when the extension finally has support for this.

Patch-set: 2
Reviewer: Hazard-SJ <464@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0