Enable Parsoid for all VisualEditor wikis
The Parsoid extension keeps the Parsoid cache service up to date on any
changes in the wiki. It is deployed in wmf6, and should hopefully also make it
into wmf5 (before this patch can be deployed). See Ide2475473c46676b9c4528ddb572376fbd0a2602.

This config only enables Parsoid for VE-enabled wikis, but after initial tests
we should enable Parsoid on all wikis the VE will launch on in July to have a
realistic load test and prime the caches.

Change-Id: I6171e0f39612d78265c689e6c11e786664b1608d