$wgEventLoggingSchemaIndexUri => $wgEventLoggingSchemaApiUri
Change I4f1763372 (to mediawiki/extensions/EventLogging.git) adds an API module
for schemas that is designed to replace 'action=raw' as the means by which
wikis fetch schema contents from MetaWiki. For the sake of clarity, the
configuration variable pointing to MetaWiki has been renamed from
'$wgEventLoggingSchemaIndexUri' to '$wgEventLoggingSchemaApiUri'. This change
prepares for the migration by leaving the old definition in tact but adding the
new one beside it.

Change-Id: I009820d1d07d7883d703c159b3d60597e253e292