Disable EventLogging and NavigationTiming on betalabs
NavigationTiming appears to be attempting to load JS modules from EventLogging, which is not fully configured on betalabs yet, and is resulting in JS errors (at least on the mobile version of the sites). In the interest of keeping betalabs working, both should be disabled until we're able to fully configure them on betalabs.
Change-Id: I82387ad1e14e2bfcd9b7e7708fd5ae0fe28edd02