Update patch set 4
Patch Set 4:

This may need to be reverted, depending on the expected lifetime of the various robots.txt files in the Varnish cache.

This change isn't supposed to go in until Google has re-indexed the pages and gotten the updated Wikipedia Zero HTML, which will *not* contain the article data, but will contain the correct per-article canonical <link>. This reindexing should be happening starting roughly at the end of the current week and I hope be complete by the middle of next week.

I believe that if Google would get a fresh robots.txt with this code (assumes the X-Subdomain header is picked up correctly), it wouldn't be able to actually re-index the pages, because it would be disallowed from doing so.

If we didn't revert, how soon would this code be deployed into production Wikipedia and how soon would the robots.txt pages in Varnish expire?

Suffice it to say this probably should have had a -2 and a bright warning to connote that it should have waited. For the moment, it seems the cached version of robots.txt is being served...not sure if that's an artifact of robots.txt actually being in the cache. I'm now wondering if a Vary: header from origin, too.

Patch-set: 4
Reviewer: Dr0ptp4kt <702@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0