beta: properly override $wgAbuseFilterCentralDB
$wgAbuseFilterIsCentral is set by the AbuseFilter extension when it is
loaded. It is done by comparing the dbname with  wgAbuseFilterCentralDB
which thus need to be set before the include.

CommonSettings-labs.php is loaded at the very end of CommonSettings.php,
that prevented the detection.  Moving the beta $wgAbuseFilterCentralDB
to InitialiseSettings-labs.php should fix the issue by propery
initializing the value before it get used.

bug: 39446
Change-Id: I224f943bcbab755779fdd0c59cd9d1cf17b46059