Enable CAPTCHA for all edits of non-confirmed users on pt.wikipedia in order to reduce editing activity
This setting was initially enabled as an emergency measure on 2008-01-25
(according to an accompanying code comment). After that, it was notified
to the community mentioning automated vandalism as a problem
(cf. <https://pt.wikipedia.org/wiki/Special:PermaLink/9164430#Activa.C3.A7.C3.A3o_de_Captcha_para_edi.C3.A7.C3.B5es_por_IP>).

Recently, this setting was disabled (cf. bug 41745). This has resulted in a
marked increase in the number of edits to the site, leading to a vote to
re-enable the CAPTCHA in order to reduce editing levels:
<https://pt.wikipedia.org/wiki/Special:Permalink/36159039>.

This change renames the "EmergencyCaptcha" global configuration variable to be
clearer and more precise.

This change partially reverts Id7b9721d9a91a30325160305b6b475ef023fd81f.

Bug: 49860
Change-Id: I3c6035ac747461f621c1698afaffec5abc3d5df4