Update patch set 1
Patch Set 1:

The Code-Review score on Gerrit should be limited to reviewing code. Please take policy questions back to Bugzilla and ptwiki. It is *perfectly* acceptable to submit a patch that implements a change that is controversial. It can be a useful way of articulating one's precise intent. If you're worried someone might merge the patch out of ignorance of the debate, note it in a comment, but don't attach a score.

Patch-set: 1
Label: Verified=0