Update patch set 2
Patch Set 2:

Limitations in Gerrit make it necessary to use Code-Review points to mark changes in this way. Comments are auto-collapsed and very easy to miss. A Code-Review -1 is highlighted in red and difficult to miss. And it seems that users can't use the other available field (Verified).

I agree with the general principle of using Code-Review scoring for code review.

Patch-set: 2
Label: Verified=0