beta: removes incubator wiki
I am not sure what the incubator wiki on beta was meant to be, but it is
no more accessible since I reduced the number of wikis on beta.  Since
the wiki is registed for autologin and is not reacheable, that produce a
console error and a missing icon whenever someone login.

Change-Id: I9021c815fcfc14eed121c6900bd061b3aa2cd43f