Update patch set 1
Patch Set 1: Code-Review-1

and would prefer it not to be available in wmf8, if possible.  in the version of Wikibase we have there, DataModel is available in a subdirectory of Wikibase.

while it probably would work, I'm not 100% sure if there would be issues with mediawiki pulling in the new extension (whatever version it's at), rather than the code in the DataModel directory. (tested and known to be okay version of code)

Patch-set: 1
Label: Code-Review=-1