(bug 50287) Restrict local uploads on Meta-Wiki
Per community request, I'm restricting local uploads on Meta
to the 'sysop' and the newly created 'uploader' user groups.

To do that, I'm taking away all upload-related user rights
from all users: 'reupload-own' which is assigned by default
to the 'user' group, and 'upload' and 'reupload' which are
assigned to the 'autoconfirmed' user group.

As requested, I'm also removing the same rights from users in
the 'confirmed' user group and creating a new 'uploader' group
with 'upload', 'reupload' and 'reupload-own' user rights. Since
no upload rights are assigned to the 'sysop' group by default,
I'm adding the said rights to that group as well (because they
wouldn't be able to upload anything otherwise).

[Unrelated: I'm removing the 'move' definition since this right
is already taken away from the 'user' group in wgGroupOverrides2.]

As requested, bureaucrats will be able to add and remove users
from the new 'uploader' group.

Bug: 50287
Change-Id: I78737791fccde75e0c90af7e6a45539d0dfdce6e