Move loginwiki to wmf9
Both for testing, and for deploying new features, it seems like
loginwiki should be on the same schedule with Phase1 wikis, so that
it always has the same or newer code than attached wikis.

There is some risk in this, in that breaking loginwiki could break SUL
for everyone. However this seems like an acceptable risk to simplify
the overall design of the SUL system.

Change-Id: I53122954b1505d2b9a83c72f0252b9c2bb4efd5a