set some paths to use $wmfHostnames 'bits' and 'test'
The asset paths were set slightly differently for beta. Since it is now
properly setting $wmfVersionNumbers to 'master' and that we have
$wmfHostnames['bits'], we can merge beta and production settings for the
following:

    $wgExtensionAssetsPath
    $wgStyleSheetPath
    $wgResourceBasePath

I have introduced $wmfHostnames['test'] pointing to either
test.wikipedia.org or test.wikimedia.beta.wmflabs.org. It is used to set
the above path settings whenever we are in the 'testwiki' context.

Thanks Anomie to have spotted this!

Change-Id: Ib0ba91f624a3b65880ae1a489f9a2c00c440de2d