Throttle now handles IP ranges.
Adding a new option parameter 'range'.

Option name ... range
Type .......... string or array
Value ......... IP range or array of such ranges

Calls MediaWiki IP class to determine if an IP belongs to a range,
or to a set of ranges.

IPv4 ("a.b.c.d/n") and IPv6 ranges are supported.

Change-Id: I2294bac734bca283b894df6c492a29f6fd18db99