(bug 50929) Remove 'visualeditor-enable' from $wgHiddenPrefs
Per community request, I'm removing the 'visualeditor-enable'
option from $wgHiddenPrefs; this will influence both enwiki
and mediawikiwiki (the only wikis with VisualEditor set to
default).

/Not hiding/ this option will enable users to opt-out from VE
using a built-in switch on the 'Editing' tab in Preferences
instead of having to hide it with a JS gadget (which does not
prevent VE from loading).

Community complaints make me assume that this option should not
be hidden again without gaining explicit community approval.

Bug: 50929
Change-Id: I885cd9a803ca3d23be72b1ebdfeaa247c3aa1d31