Update patch set 2
Patch Set 2: Code-Review-1

this is not the issue.

The problem is that these permissions are defined in Wikibase, with defaults set for anons '*'

Wikibase gets loaded after groupOverrides are applied.

The solution is either to have groupOverrides happen after extensions load or in this specific case, define this setting for testwikidatawiki in CommonSettings.php where we define wikidata settings.

Patch-set: 2
Reviewer: Aude <33@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1