Disable "Mark as helpful" extension on English Wikipedia.
This extension is used by MoodBar to enable a user to mark a comment
as helpful. When present, it adds JavaScript code that uses a CSS
selector to dynamically add a "Mark this response as helpful" link
below help messages left on user talk pages via MoodBar.

When the extension is absent, those will just show up as whitespace.

Since MoodBar itself was disabled in February 2013 in commit
57984050217e4ac0e4db4e2ef9321f3686a2a0eb, there is no point in
retaining its dependency at this point.

Change-Id: Ida7d70daadb9d0867cab24caef252b5facda55df