Update patch set 4
Patch Set 4:

We talked on the thread how that preference should not be configured in operations/mediawiki-config but in the extension itself.

If visualeditor-betatempdisable user preference is removed from Visual Editor, this patch is not going to make it working anyway since the preference will not be there anymore.

Patch-set: 4
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0