Update patch set 2
Patch Set 2:

Is it good or bad to have the bug number in the oneliner? Some people do it, some do not. I once had someone come along and add it when I did not include it, and now it is being taken out when I put it in. Quite frustrating.

Patch-set: 2
Reviewer: TTO <644@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0