Update patch set 14
Patch Set 14: Code-Review-1

I suggest changing the variable name in a separate commit. The discussion about the variable name is a distraction, and makes deployment more difficult.

Also, I don't think the word "temporary" should appear in the variable name. There's no other temporary setting that is described as such, and there's no guarantee that the next time a general edit captcha is enabled, it will be temporary. For technical reasons, it's better if variable names are not part of the policy battlefield.

Patch-set: 14
Label: Code-Review=-1