Send rel=canonical pointing from HTTPS to HTTP
Bug 48402: Google has been randomly indexing HTTPS pages in preference
to HTTP, which has caused the load on the SSL cluster to increase to
near capacity. If and when we switch completely to HTTPS, we want to be
able to control that migration, rather than having Google do it for us.

Also set wgCanonicalServer to HTTPS for all wikis that are
unconditionally redirected to HTTPS by Apache configuration. It probably
should have been done when the redirects were set up, but enabling
rel=canonical makes it actively harmful to redirect from canonical to
non-canonical URLs.

Change-Id: Ic5ba90864af4611f2aeae4afdf802b3966b1aa7a