Give testwiki some custom namespaces
Adding Portal: (common Wikipedia namespace) and
Index: (common Wiktionary namespace) to testwiki.
No special functionality - they are just for tests.

I didn't use the Book name, as suggested, because
that ought to be hooked up to the Collection extension,
and doing so would be unnecessary on testwiki
(testing of Collection is usually done on test2wiki).

Bug: 52528
Change-Id: I94c34799c70a6d471cd113602a634527171ea40e