Update patch set 2
Patch Set 2:

TL;DR: enwiki wanted to get rid of large-scale AFTv5 deployment, but could still have AFT on a small minority. dewiki wanted to get rid of it entirely.

If anything seems incorrect there, please let me know - this is how I've understood it.


For enwiki:

Following http://en.wikipedia.org/wiki/Wikipedia:Requests_for_comment/Article_feedback, AFTv5 was removed from most of the pages it was enabled on. It had been on 10% + articles with Category:Article_Feedback_5 & Category:Article_Feedback_5_Additional_Articles; the categories have been removed & lottery percentage set to 0.

A "fair sized minority" did support the tool and it could live on "limited to a level which that minority cope with on its own". While the large-scale deployment of AFTv5 was undone, it is still possible for individuals to opt in (http://www.mediawiki.org/wiki/Article_feedback/Version_5/Feature_Requirements#Enable.2Fdisable_feedback_on_a_page) to AFTv5 for certain pages.

It appears to be enabled on ~2800 pages now: http://ee-prototype.wmflabs.org/metrics/aftv5_enabled_pages.php?wiki=enwiki


For dewiki:

Category-based & lottery have both been shut down completely.

They didn't want to leave a possibility for editors to enable AFTv5 on specific pages, but did not yet completely want to remove AFTv5, to cope with existing feedback.

This is what this patch was for: it'd disable the opt-in feature that allowed individual editors to enable AFTv5 on individual pages (and which still is possible on enwiki)

Patch-set: 2