Dereference unused category from ArticleFeedbackToolv5 en.wiki config
The categories are empty since August 14 as they are deprecated.
They are replaced by the new JS switch available on all pages.
<https://en.wikipedia.org/w/index.php?title=Wikipedia_talk:Article_Feedback_Tool/Version_5&oldid=569251758#Re-enabled_on_all_pages>

I preferred categories, but if they are unused this should be
uncontroversial cleanup of stale configuration, not to mention how
confusing it is to have two concurring systems to control AFT.

Change-Id: Iad069551c360d2c8cc6e7ba8477abfeff386414e