Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Let's try it!  Note that the Flow team has not promised any DB compatibility -- "All those comments will be lost in time, like tears in rain"

My patch comment is a way to enable Parsoid in the right situation, but it needs review. It can be a different commit.

Patch-set: 2