Update how VisualEditor's experimental config works
Previously, we used $wmgVisualEditorExperimental to decide whether a
wiki would have VisualEditor's "experimental" code enabled for all,
or disabled for all. With the forthcoming "BetaFeatures" system, we
won't be able to set it in the old way:

  $wgVisualEditorEnableExperimentalCode = true;

… and instead will use a user preference that users can opt-in to
pick on a per-user basis. Thus, to set this for all users on a given
wiki, we need instead to use:

  $wgDefaultUserOptions['visualeditor-enable-experimental'] = 1;

Change-Id: Id1a977ee8364b0403b5f86c9f0337f17920e1b0d