Remove EditPageTracking extension
The extension appears to be unused: it has had no substantial commits since
2011 and has not been updated to use EventLogging rather than the production
databases. It logs the first edit page impression for users that registered
after July 2011. It was presumably set to this value to ensure that data was
only collected from newly-enrolled users, but the passage of two years has made
the configuration nonsensical and the data it collects worthless. Having a
fixed and unmaintain cut-off date also means that the database load imposed by
the extension gradually swells, too. Let's get rid of it.

Change-Id: Idcb034791ade61fc5268675945a6c9d4d4864dba