Update patch set 2
Patch Set 2:

We definitely want this in beta in the future, since I've found several bugs that happen only when wgSecureLogin is enabled. But if it's currently preventing other tests from running, I'm fine with this going in, and we'll revert this as soon as we have an official cert.

Patch-set: 2
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0