(bug 4883) Edit $wgSiteName and set up NS aliases for ukwikinews
This patch changes $wgSiteName and $wgMetaNamespace for ukwikinews
from the current 'ВікіНовини' into 'Вікіновини', and changes the name
of NS_PROJECT_TALK accordingly, while keeping the current names
of these two namespaces as aliases for backward compatibility.

I'm also adding new aliases for namespaces 12, 14 and 104, and another
one for NS_PROJECT (NS 4) per community request. See bug for URL
with on-wiki consensus for these changes.

Bug: 48843
Change-Id: I4429fb327159b6e71a30c6536ccab0fcf60e6f66