Add templateeditor right, group, and restriction
For en.wikipedia; this change adds the 'templateeditor' restriction
level 'between' the 'autoconfirmed' and 'sysop' levels intended for
appying to high-traffic templates that are currently protected.

This commits creates the group 'templateeditor' that has a right
with the same name, allowing members to edit the protected templates,
and gives that right (as well as the ability to grant and revoke
the group) to the 'sysop' group.

Bug: 55432

Change-Id: I2c2bdf0934de46ae118d22ad0260a90d02d1375b