Update patch set 2
Patch Set 2:

Yeah, this is easy to revert if it causes unexpected consequences (like I said at the bug, not tested due to a lack of a suitable environment) or when the SSL works.

But from my point of view, we should try to have working SSL (i.e. the browser trusts it without ignoring bad certs, and preferably out of the box), or no SSL.

Patch-set: 2
Reviewer: Mattflaschen <519@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0