(bug 48480) Remove EmailCapture extension settings
This patch removes the settings for the unused EmailCapture
extension from CommonSettings.php and InitialiseSettings.php.
The extension does not seem to have been ever deployed to the
cluster, and it is only used by AFTv4, which is disabled
globally anyway; see bug for more information on the subject.

I'm also using this opportunity to remove a whitespace I left
behind yesterday :-)

Bug: 48480
Change-Id: Idbcf67cf1b3ebca840bd35fb70266e678750c442