(bug 54828) Enable FlaggedRevs for Portuguese Wikipedia
This patch provides configuration for the FlaggedRevs
extension for the Portuguese Wikipedia and adds it to
flaggedrevs.dblist per InitialiseSettings.php instructions.

The exact configuration has been kindly provided by
Danilo.mac; see bug for URL to its on-wiki version and to
a page that lists community consensus for enabling the
extension.

Bug: 54828
Change-Id: I868cbb822f39ce28c7dd404ab7a8f0011255fee9