multiversion/checkoutMediaWiki: create/update 'current' / 'stable' symlinks
Bits static asset URLs currently reference a specific production branch. This
is fine in most cases, but in certain cases (such as large files in extensions
that change infrequently) we want a stable URL to the latest version of some
asset, so that the referenced asset is not thrown out of the browser cache on
each release. The current case is the UniversalLanguageSelector extension,
which serves large font files out of its repository.

This patch updates the multiversion/checkoutMediaWiki script so that in
addition to checking out a new production branch, it also updates a pair of
directories in the Bits docroot, static-current & static-stable, that contain
symlinks to the asset paths of the most recent & secondmost recent production
branches, respectively. If only one production branch exists, both current &
stable pointers are pointed to it.

Bug: 55506
Change-Id: If597d70625cf8f06c0e59dd7749442ba05769174