Enable VisualEditor for NS_FILE, NS_HELP, NS_CATEGORY
VisualEditor should work fine in these three namespaces, as they are not
signed and do not have fragments of partial wikitext.

TODO: Should this also extend to all $wmContentNamespaces for wikis that
have that? What about Portal, which isn't set as such on enwiki but is?

Also clean up MWwiki's VE namespace definition to extend, not replace.

Bug: 55968
Change-Id: I2d378203b4459fd459653a18ca4a060c43086a65