Update patch set 2
Patch Set 2:

I don't care which account name is used. I'm happy to surrender "EdwardsBot" if that's deemed the best choice. My only related thought is that, whichever name is used, we should be explicit about it in the Wikimedia configuration.

I'm excited and eager to get this tool deployed to the rest of the Wikimedia wikis. :-)

Patch-set: 2
Label: Verified=0