Update patch set 2
Patch Set 2:

Did we test that this won't also make the File:video.webm pages on commons do a pop up? I don't think we want that. 

I think this is positively received for small in-article videos. But may need to think how it would work for file pages, and if the article has a focus on a video giving it over 400px, I think playing inline with normal fullscreen would make more sense then a pop up that does not change the size of the embed much.  ... 

Maybe the check could be something like if video max true pixel size is > embed size and embed is < 400 px.

Patch-set: 2
Reviewer: Mdale <199@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0