Enable Flow discussions on a few test wiki pages
Only testwiki, test2wiki, and mediawiki.org, and only on two or three
pages on each: 'Talk:Flow QA', 'Talk:Sandbox' (and 'Talk:Flow' on
mw.org).

Configure Flow to store in HTML and to use a cross-wiki DB and
$wgDefaultExternalStore for most wikis.

Move Flow from wmf-config/extension-list-labs to the main
wmf-config/extension-list.

Bug: 56506
Change-Id: I1fba27a98059686056e82e15ab5f749a6eea14b1