Update patch set 3
Patch Set 3:

Meh, don't have to symlink. But with all the cleanups it'll be way easier to just drop these in as-is.

With the latest revisions to it, the only part that would really conflict would be Cdb.php on the exception still (but we still might come up with some clever way around that).

Patch-set: 3
Label: Verified=0