Add a githook for logging repo modification on tin
The SAL would be a lot more useful if it contained information about what went
out with each deployment. Although there is a convention around document
deployment payloads in log messages, it is not consistently followed.
So: automate.

This patch adds a script, 'logmsg-git-hook'. When invoked with '--install', it
configures itself as a handler for post-commit, post-merge, and
post-rewrite git hooks for /a/common & its php-* subdirs.

When invoked as a githook, it !logs the Change-Id of the latest non-merge
commit shared between the local branch and upstream. These criteria are meant
to ensure that we don't !log security patches by mistake.

Change-Id: I9981357261f607dbddaea247dbe8439af81d4fc9