Update patch set 1
Patch Set 1:

There are two possibilities I would sugget.

[ First possibility ]

(1) you read the config files to get a better understanding of this wmf/wmg feature, or ask questions to the relevant people and then document it clearly on a wiki page, so we all know about this

(2) you then use this understanding to cleverly select the good variables to migrate.

This method would have the big advantage the community as a whole would have a reference about what variable should have a $wmg prefix, what variable should have a $wmf prefix.

[ Second possibility ]

Now, if I would understand if you don't want to do an analysis/documentation effort and are more here here to give a hand on a specific config migration need.

In this case, yes, you should only migrate the ones indicated in bug title and comment 2.

Patch-set: 1