Update patch set 1
Patch Set 1: Code-Review+1

I approve, and I'm ok with this going out now as if it breaks we should see it pretty quickly (and it's an easy revert).

But, if it breaks during the normal deploy week, we might negatively effect our ability to diagnose a performance issue in the moment.

Patch-set: 1
Reviewer: Greg Grossmeier <680@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1