Mobile m. and zero. landing page redirect handling by ZERO
This script is similar to extract2.php - it handles redirection
of m.wikipedia.org and zero.wikipedia.org to the specific language
version depending on the zero configuration.

For non-zero customers, it keeps current logic of redirecting to EN

Soon zero will be able to redirect to sister projects as well,
but it will be based internally on the HOST header.

Change-Id: I58b03cf6751b2791c2daddf7800889d13ab421f9