Update patch set 2
Patch Set 2: Code-Review-1

I think at the time of writing this, we still thought real Commons is configured as a ForeignAPIRepo. Actually it is a ForeignDBViaLBRepo, so it makes more sense to use the same configuration for Beta Commons, and that already happens (in filebackend-labs.php).

Patch-set: 2
Reviewer: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1