Update patch set 3
Patch Set 3: Code-Review-1

Do not +2 until we have our window.

Is $wgFlowExternalStore = $wgDefaultExternalStore; good enough? I think it defaults to empty if not set up.

My understanding is in production we want Flow using a cross-wiki DB (named e.g. 'flow') on the extension1 cluster. The former is $wgFlowDefaultWikiDb , but I don't see the equivalent of $wgArticleFeedbackv5Cluster or $wgEchoCluster in the Flow code.

Patch-set: 3
Label: Code-Review=-1