Update patch set 1
Patch Set 1: Code-Review-1

Like I said on the bug, I think there are unanswered questions regarding functionality and permissions for this namespace that are (for now) blockers.

If we can settle the issues I brought up on the bug regarding who can create/publish drafts and how, then I think we can move forward with this. There is additional functionality we'd like to experiment with in the future, but some of that can be built in after a basic namespace is created.

Patch-set: 1
Reviewer: Swalling <468@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1