Update patch set 1
Patch Set 1:

The only unanswered question I see is that of permissions, and the obvious answer there is that anyone should be allowed to create pages in this namespace. Perhaps this needs to be put to the community again.

But that is all. It seems to me that any other "unanswered" questions are supplementary to this configuration change, not blocking it.

As far as functionality is concerned, the community has done a fine job of writing AfC review scripts in the past, and there is no reason why these scripts could not be modified (or new ones written) to allow this namespace to come into use very soon.

Do we need to give the community time to nut out the relevant policy? Maybe so, but that hasn't blocked configuration changes in the past.

Patch-set: 1
Label: Verified=0