Update patch set 1
Patch Set 1:

re: "The only unanswered question I see is that of permissions, and the obvious answer there is that anyone should be allowed to create pages in this namespace. Perhaps this needs to be put to the community again."

This question was asked of the RFC closer. Tom will reply on Bugzilla and clear the issue up. I assume that we want to let anons create pages in Draft, not Draft talk, but his comments would be helpful. 

The permissions issue is not just page creation, but also publishing to mainspace.

Patch-set: 1