Update patch set 1
Patch Set 1:

Anon creation of drafts is one of the more fundamental things that we should be adding after the fact in some vague way.

The creator of the proposal and other community members involved in the process on enwiki have added to the bug that they would like to wait, and do this right from the beginning with clear agreement about the basic requirements, rather than rushing in to things.

Patch-set: 1