Update patch set 2
Patch Set 2:

We also need to add the namespace to wgExemptFromUserRobotsControl so people can not force indexing manually using the __INDEX__ magic word.

Allowing that magic word here doesn't make sense; if it's ready to be indexed, it should be published to the main namespace.

Note that wgExemptFromUserRobotsControl not overridden in mediawiki-config yet.  The default value of wgExemptFromUserRobotsControl is wgContentNamespaces.  However, I don't think we want the draft namespace to be content.

So wgExemptFromUserRobotsControl should be set to wgContentNamespaces + draft_namespace (the draft namespace id can be passed to CommonSettings via a wmg).

Patch-set: 2