Update patch set 6
Patch Set 6: Code-Review-1

Jenkins fails the php lint because the symbolic link of mobileredirect.php points to a full path which is not on the box.

That could be safely ignored.

BUT, the symbolic link redirect points to /w/mobileredirect.php which DOES NOT exist.

Patch-set: 6
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1