Update patch set 1
Patch Set 1:

Dan, amend this changeset and put your settings in the new feature flag section that Mark made in wmf-config/CommonSettings-labs.php.

If you need help figuring out the specific configuration values I would suggest talking to Antoine (hashar on irc) and/or Aaron to get some hints as well as looking at the rest of the config in the operations/mediawiki-config repository.

Patch-set: 1