Re-add the docroot/secure directory
It is still used by the secure.wikimedia.org vhost. It was apparently
deleted accidentally in I3e12805cb, when the last file in the directory
was removed. Git does not preserve empty directories.

I added a 404.html file which I will reference in a subsequent commit to
the apache configuration. Currently, missing files give a 301 redirect
to https://meta.wikimedia.org/w/404.php

Change-Id: Id5b4f431da2ff6d1ecf6279be44820f2e0781227