Update patch set 4
Patch Set 4:

(1 comment)

Legoktm, it's no longer an issue since MZMcBride changed the patch to "noindex,nofollow". However, http://www.robotstxt.org/faq/relnofollow.html indicates the link nofollow and meta nofollow are different ("From that description it sounds like it only affects the ranking, and the Google robot may still follow the links and index them. If so, it is different from the robots meta tag NOFOLLOW semantics.").  We should have both for clarity.

Gating this on enwiki specifically isn't ideal.  However, we can clean that up if/when we generalize this to another wiki.

I was originally thinking about wgExemptFromUserRobotsControl just for Draft (not Draft_talk), but we can revisit that later if need be.

The rest looks okay, so I've created I3520527afd48722d09f57dc959402184a5d59600 to test it on Labs.

Patch-set: 4