Update patch set 3
Patch Set 3:

Chad and I had a talk and figured this'd be the simplest way to maintain which wikis don't get cirrus as a BetaFeature.

Please check my wmg-foo.  I'm not 100% sure what black magic backs this configuration system and I really don't have a way to test this locally.

Patch-set: 3
Label: Verified=0