Enable Cirrus as a beta feature
Add $wmgCirrusIsBuilding which we'll set on all wikis that are currently
building their CirrusSearch index for the first time.  If the wiki doesn't
have that but has Cirrus as a secondary then Cirrus will be available as
a BetaFeature.

The plan:
When adding new wikis, set this to true for them.  When they are done
indexing set it to false.  After that, announce to whoever want to know
that Cirrus is available as a beta feature.

Change-Id: Ia7c593fb730f96cc6228585892d743cce78a8763