Update patch set 3
Patch Set 3:

CC-ing Isarra, the author of the new Wikipedia favicon so that she can have a look at the new icon, as there is some difference between her version and the one uploaded by Mateusz.

See <http://tools.wikimedia.pl/~odder/favs/> and compare the -new icons with the -old ones; it is my impression that the old icon is sharper than the new one.

That said, as long as Isarra approves of the new icon, I'm OK with it as well.

Patch-set: 3