Update patch set 1
Patch Set 1: Code-Review+1

$wgCopyUploadsDomains is the whitelist for UploadFromUrl::isAllowedHost(). If the list is empty then all hosts are allowed.

I'm in favor of this change because it will allow GWToolset to function as designed, but it's a big change and I have no idea who would have to approve for this to actually be carried forward into production.

Hopefully some of the folks I added to the review will be able to chime in with learned answers.

Patch-set: 1
Label: Code-Review=+1