Update patch set 3
Patch Set 3:

I was unable to find a higher resolution version of the internal.ico favicon, and it was difficult to increase the size of the icon to 32px and 48px without it becoming even more blurry, so I improvised by scaling the current icon up, and tracing over it. I'll attempt to find a higher resolution version and/or fix the current version.

Patch-set: 3
Label: Verified=0