Update patch set 3
Patch Set 3:

Odder: to be fair, 3 days over the weekend is not that long. I don't think it is going to be too disruptive for them :-] After all the favicon still looks the same.

Tholam: feel free to propose a new  logo, might want to propose your sketch to that community. I am sure they will love whatever you can come with.

Patch-set: 3