Enable Flow discussions on a few wikis' test pages
Only mediawiki.org, testwiki, and test2wiki, and only on two or three
pages on each: 'Talk:Flow QA', 'Talk:Sandbox' (and 'Talk:Flow' on
mw.org).

Configure Flow to store in HTML and to use a cross-wiki separate DB on
extension1 cluster and $wgDefaultExternalStore for most wikis.

Adding Flow to extension-list is a separate change 100551

Bug: 56506
Change-Id: I1fba27a98059686056e82e15ab5f749a6eea14b1