Cirrus as secondary for more wikis
Cirrus as secondary for all wikimanias, wikimedias, and wiktionaries.

We must carefully build the list of $wmgCirrusIsBuilding because we don't
want to include those that have already been built or else they will lose
the ability to set cirrus as a beta feature.

Also, wikidata gets Cirrus as a beta feature now that it is done indexing.

Finally, wmgUseCirrus overrides wmgUseCirrusAsAlternative, not the other
way around.  This will let us add wikis that want cirrus as primary to
cirrus.dblist while adding dblists to wmgUseCirrusAsAlternative.

Change-Id: Iec7932f07e2095da8fa38497e7a89bf3501de5aa