Update patch set 5
Patch Set 5:

Expi1: If you have a look at the favicon in a graphic manipulation program (such as The GIMP or Photoshop), you will notice that it includes a white rectangle below the Wikimedia logo, on which the letters 'inter' are visible in black color. I think it could be possible to replicate that.

Patch-set: 5