Update patch set 12
Patch Set 12:

I tried to make the commit message more clear.  the "Wikidata" git repo loads all the dependencies as submodules and with a "Wikidata" entry point.  The goal is to eventually have one git repo that contains all the wikidata extensions + libs, we think will be easier to manage that way.

Patch-set: 12
Label: Verified=0