(bug 55630) $wgCategoryCollation = 'xx-uca-ckb' for ckbwiki
To be deployed only after I0b745d95 and I7d7f0075 in core are deployed.

This does not require updateCollation.php to be ran (internally the
same collation is used), but that probably should still be done to
keep the data in the tables sane (and the wiki is small anyway).

Change-Id: Ie8036cd2d3898d17f6904a8ed6ebef1c3f8d3bda