Cirrus config updates
* Set cirrus as secondary on all wikisources and commonswiki.
* Let all wiktionaries, wikimania, and wikimedia wikis have cirrus as a
beta feature.
* Rename $wgCirrusSearchContentReplicaCount to $wmgCirrusSearchReplicaCount
* Add support for setting $wgCirrusSearchNamespaceMappings and set it up
for commons.
* Add support for setting $wgCirrusSearchReplicaCount because it is required
for $wgCirrusSearchNamespaceMappings and set it for commons.
* Add support for settings $wgCirrusSearchShardCount because it is required
for $wgCirrusSearchNamespaceMappings and because we would like to change
the number of shards from Cirrus's defaults.
* Set number of shards for all wikis.  The default is a single shard.
Larger wikis will get more with the goal of each shard being about 2GB
with a bit of room to grow.
* Add warning to $wgContentNamespaces that changing it for Cirrus wikis
will remove the affected namespaces from search results until an full
reindex is completed.

DEPLOYMENT: This change requires the cirrus indexes be built for all wikis
listed until $wmgCirrusIsBuilding in intialize settings.

DEPLOYMENT: This change allows us to do an in place rebuild of all currently
built Cirrus indexes to reduce the number of shards they use which should
reduce load on the Elasticsearch cluster, improve recovery times when a
node drops out, and potentially speed up search and indexing.

Change-Id: If5e0f1bb4bb571a2ae5bb4bedfb52a2a7a10f706